Improved normalizing algorithm
Mask files are now interpreted more intelligently when also normalizing and
cropping. Instead of throwing away mask *positions* that lie outside the
cropping area, we now only throw away masks that do not *overlap* with the
cropped field of view. This way one can use a 'course' mask file with
subapertures that can be slightly too big, but still crop a smaller field
out of the subimage.